Add files via upload
created the Tkinter version of the college website, you can extract student details by entering the username and password using this Tkinter application. The option to see marksheet will be available soon in tkinter. And the UI will be also improved in the next versions !!!

This tkinter version can be run without any browser with an internet connection, it's windows installer package will also be available in next few commits...